+++ modules/@apostrophecms/ui/ui/apos/apps/AposBreakpointPreview.js
@@ -1,0 +1,5 @@
+import { useBreakpointPreviewStore } from 'Modules/@apostrophecms/ui/stores/breakpointPreview.js';
+
+export default function() {
+  useBreakpointPreviewStore();
+};