+++ modules/@apostrophecms/widget-type/ui/apos/components/AposWidgetEditor.vue
@@ -271,7 +271,10 @@ export default {
         return;
       }
       if (this.preview.create) {
-        this.preview.area.remove(this.getPreviewWidgetIndex(), { autosave: false });
+        this.preview.area.remove(
+          { index: this.getPreviewWidgetIndex() },
+          { autosave: false }
+        );
       } else if (!this.saving) {
         this.preview.area.update(this.previewSnapshot, {
           autosave: false,